<feat>(project): 兼容TLC59108灯条 版本更新至V1.2.3 未实测 修改点如下:
1.gd32f3x0_it.c中：SysTick_Handler()中增加TLC59108灯条渲染任务
2.register_info.c中：
  <*> 为兼容以前的TLC59108灯条驱动寄存器配置 调整寄存器定义
  <*> 根据调整后的寄存器定义 修改相关control逻辑
3.board.c中：修改灯带默认参数初始化的参数以及接口定义
4.drv_pwm.c中：
  <*> 为兼容TLC59108的pwm周期设置 调整_pwm_timer_init()中psc参数设置
  <*> _control_pwm()方法中优化对驱动类型和驱动设备重新初始化方法的逻辑处理
5.led_bar.c中：
  <*> 增加TLC59108灯条对象 使用Lbar作为统一对象管理
  <*> 优化init_led_bars()初始化灯条方法中初始化逻辑 并增加对TLC59108灯条的初始化
  <*> 调整led_bar_control()灯条控制方法中解析逻辑 并增加对TLC59108灯条设备的处理
  <*> 实现TLC59108灯条的dimming/blinking方法
6.led_frame.c中：
  <*> on()和off()方法中 增加对灯条set_color()设置颜色的回调函数有效性判断
7.task_sch.c中：
  <*> 调整原任务等级中任务时间的变量类型为uint32_t
  <*> 对直接的一些内置方法进行接口说明
8.工程文件夹中增加bsp文件夹 用于对WS2812/TLC59108灯条的相关文件管理
9.增加tlc59108.c/tlc59108.h tlc59108_led_bar.c/tlc59108_led_bar.h文件
10.common.h中：修改dev_ops结构体属性中control()方法的cmd参数类型为const int
11.drv_pwm.h中：修改原TLC59108灯条PWM波驱动频率相关的宏定义 时速需满足100KHz以上
12.led_bar.h中：
  <*> 为兼容TLC59108灯条数据手册中对dimming/blinking模式的数值定义 修改相关枚举宏数值
  <*> 调整driver_recv_len枚举类型中 两种灯条写入的字节数
  <*> 优化BAR_REQ_LEN_CHECK()参数有效性判断方法
  <*> 新增版本至V1.2.3 并写明版本注释
13.register_info.h中：
  <*> 调整两种灯条寄存器参数区偏移地址
  <*> 调整两种灯条工作模式寄存器地址
  <*> 优化CTRL_REG_CMD_CHECK()参数有效性判断方法

+++ Template/platform/Inc/task_sch.h
@@ -87,7 +87,7 @@ void task_register(const TaskType task_type,
  */
 Rtv_Status task_ms_reset(const TaskType task_type,
                          const TaskLevelType task_level,
-                         const uint16_t _new_task_ms);
+                         const uint32_t _new_task_ms);
 
 /**
  * @brief 任务调度接口额
